Module parachain_template_node::dev_service

The development service provides a way to run a parachain runtime in a real blockchain context, but without a backing relay chain. This allows developers to quickly and easily spin up parachain nodes using the –dev flag, for example. It can also be used in integration tests.

Structs

ExecutorDispatch

NativeElseWasmExecutor

A generic CodeExecutor implementation that uses a delegate to determine wasm code equivalence and dispatch to native code when possible, falling back on WasmExecutor when not.

Functions

new_dev

Builds a new development service. This service uses manual seal, and mocks the parachain inherent.

new_partial

Type Aliases

FullBackend 🔒

FullClient 🔒

FullSelectChain 🔒