Function parachain_template_node::service::new_partial

pub fn new_partial(
    config: &Configuration
) -> Result<PartialComponents<TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<ParachainNativeExecutor>>, TFullBackend<Block>, (), DefaultImportQueue<Block>, FullPool<Block, TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<ParachainNativeExecutor>>>, (ParachainBlockImport<Block, Arc<TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<ParachainNativeExecutor>>>, TFullBackend<Block>>, Option<Telemetry>, Option<TelemetryWorkerHandle>)>, Error>

Starts a ServiceBuilder for a full service.

Use this macro if you don’t actually need the full service, but just the builder in order to be able to perform chain operations.